dependency visualizer takes manifest as props
