+++ packages/app/src/components/DependencyVisualizer/DependencyVisualizer.tsx
@@ -1,14 +1,9 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useSearchParams } from "react-router";
-import {
-  getAuditManifest,
-  getDependencyManifest,
-} from "../../service/api/index.ts";
 import type { AuditManifest, DependencyManifest } from "@napi/shared";
 import { SidebarProvider, SidebarTrigger } from "../shadcn/Sidebar.tsx";
 import { Button } from "../shadcn/Button.tsx";
 import { Moon, Sun } from "lucide-react";
-import { toast } from "sonner";
 import { useTheme } from "../../contexts/ThemeProvider.tsx";
 import { FileExplorerSidebar } from "./components/FileExplorerSidebar.tsx";
 import BreadcrumbNav from "./components/BreadcrumNav.tsx";
@@ -17,7 +12,6 @@ import FileVisualizer from "./visualizers/FileVisualizer.tsx";
 import SymbolVisualizer from "./visualizers/SymbolVisualizer.tsx";
 
 export interface VisualizerContext {
-  busy: boolean;
   dependencyManifest: DependencyManifest;
   auditManifest: AuditManifest;
   highlightedCytoscapeRef: {
@@ -26,17 +20,13 @@ export interface VisualizerContext {
   } | undefined;
 }
 
-export default function DependencyVisualizer() {
+export default function DependencyVisualizer(props: {
+  dependencyManifest: DependencyManifest;
+  auditManifest: AuditManifest;
+}) {
   const [searchParams] = useSearchParams();
 
   const { theme, setTheme } = useTheme();
-
-  const [busy, setBusy] = useState<boolean>(true);
-
-  const [auditManifest, setAuditManifest] = useState<AuditManifest>({});
-  const [dependencyManifest, setDependencyManifest] = useState<
-    DependencyManifest
-  >({});
 
   const [highlightedCytoscapeRef, setHighlightedCytoscapeRef] = useState<
     {
@@ -45,36 +35,6 @@ export default function DependencyVisualizer() {
     } | undefined
   >(undefined);
 
-  useEffect(() => {
-    async function handleOnLoad() {
-      setBusy(true);
-      try {
-        const dependencyManifestPromise = getDependencyManifest();
-        const auditManifestPromise = getAuditManifest();
-
-        const allPromise = Promise.all([
-          dependencyManifestPromise,
-          auditManifestPromise,
-        ]);
-
-        toast.promise(allPromise, {
-          loading: "Loading manifests",
-          success: "Manifests loaded successfully",
-          error: "Failed to load manifests",
-        });
-
-        const [dependencyManifest, auditManifest] = await allPromise;
-
-        setDependencyManifest(dependencyManifest);
-        setAuditManifest(auditManifest);
-      } finally {
-        setBusy(false);
-      }
-    }
-
-    handleOnLoad();
-  }, []);
-
   return (
     <SidebarProvider
       defaultOpen
@@ -82,9 +42,8 @@ export default function DependencyVisualizer() {
       style={{ "--sidebar-width": "30rem" } as React.CSSProperties}
     >
       <FileExplorerSidebar
-        busy={busy}
-        dependencyManifest={dependencyManifest}
-        auditManifest={auditManifest}
+        dependencyManifest={props.dependencyManifest}
+        auditManifest={props.auditManifest}
         onHighlightInCytoscape={(node) => {
           if (!node.fileId) return;
           const newRef = {
@@ -160,29 +119,26 @@ export default function DependencyVisualizer() {
           {searchParams.get("fileId") && searchParams.get("instanceId")
             ? (
               <SymbolVisualizer
-                busy={busy}
                 fileId={searchParams.get("fileId")!}
                 instanceId={searchParams.get("instanceId")!}
-                dependencyManifest={dependencyManifest}
-                auditManifest={auditManifest}
+                dependencyManifest={props.dependencyManifest}
+                auditManifest={props.auditManifest}
                 highlightedCytoscapeRef={highlightedCytoscapeRef}
               />
             )
             : searchParams.get("fileId")
             ? (
               <FileVisualizer
-                busy={busy}
                 fileId={searchParams.get("fileId")!}
-                dependencyManifest={dependencyManifest}
-                auditManifest={auditManifest}
+                dependencyManifest={props.dependencyManifest}
+                auditManifest={props.auditManifest}
                 highlightedCytoscapeRef={highlightedCytoscapeRef}
               />
             )
             : (
               <ProjectVisualizer
-                busy={busy}
-                dependencyManifest={dependencyManifest}
-                auditManifest={auditManifest}
+                dependencyManifest={props.dependencyManifest}
+                auditManifest={props.auditManifest}
                 highlightedCytoscapeRef={highlightedCytoscapeRef}
               />
             )}
