+++ packages/app/src/components/DependencyVisualizer/visualizers/SymbolVisualizer.tsx
@@ -165,12 +165,12 @@ export default function SymbolVisualizer(
 
       <div className="absolute bottom-10 left-1/2 transform -translate-x-1/2 z-20">
         <Controls
-          busy={props.busy || busy}
+          busy={busy}
           cy={symbolVisualizer?.cy}
           onLayout={() => symbolVisualizer?.layoutGraph(symbolVisualizer.cy)}
         >
           <GraphDepthExtension
-            busy={props.busy || busy}
+            busy={busy}
             dependencyState={{
               depth: dependencyDepth,
               setDepth: handleDependencyDepthChange,
